requirements.txt
streamlit
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
neo4j
transformers==4.41.2
python-dotenv
wikipedia